Add pack function, move functions from base.schemas to base.helpers
